requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
requests==2.32.3
torch==2.2.2
transformers==4.48.3
tokenizers==0.21.0
huggingface-hub==0.28.1
numpy==2.2.2
groq==0.18.0
python-dotenv
